Sorting arbitrary constants should not be done, as it relies on `DefId` ordering, which breaks incremental compilation.
